# docs/partners/apex/openapi.yaml
openapi: 3.0.3
info:
  title: Apex Logistics Integration API
  version: 1.0.0
  description: Synthetic REST/JSON contract for Apex Logistics in the FreightBridge portfolio lab.
servers:
  - url: https://api.apex-logistics.example.com
    description: Placeholder production endpoint
  - url: https://sandbox.apex-logistics.example.com
    description: Placeholder sandbox endpoint
security:
  - bearerAuth: []
paths:
  /v1/load-tenders:
    post:
      summary: Create and tender a load inside Apex
      description: Apex-owned simulator/business action that creates or tenders a load within Apex Logistics. This is not the future FreightBridge inbound endpoint. Apex outbound delivery to FreightBridge is documented separately in outbound-load-tender-contract.md.
      operationId: createApexLoadTender
      tags: [Load Tenders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApexLoad'
            examples:
              validLoad:
                $ref: '#/components/examples/ValidLoadTender'
      responses:
        '202':
          description: Apex load tender created or queued inside Apex
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedLoadTender'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessValidationFailure'
        '500':
          $ref: '#/components/responses/SystemFailure'
        '503':
          $ref: '#/components/responses/DependencyFailure'
  /v1/tender-responses:
    post:
      summary: Receive a tender decision
      description: Apex-owned endpoint that will eventually receive transformed tender decisions from FreightBridge.
      operationId: receiveTenderResponse
      tags: [Tender Responses]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApexTenderResponse'
            examples:
              accepted:
                $ref: '#/components/examples/TenderAccepted'
              rejected:
                $ref: '#/components/examples/TenderRejected'
      responses:
        '202':
          description: Tender response accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedPartnerEvent'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/LoadNotFound'
        '422':
          $ref: '#/components/responses/BusinessValidationFailure'
        '500':
          $ref: '#/components/responses/SystemFailure'
        '503':
          $ref: '#/components/responses/DependencyFailure'
  /v1/load-tenders/{loadId}/dispatch:
    post:
      summary: Dispatch an Apex load tender to FreightBridge
      description: Explicit Apex simulator action that sends an existing ApexLoad to the FreightBridge inbound Apex integration endpoint. This does not contact Midwest or generate X12.
      operationId: dispatchApexLoadTender
      tags: [Load Tenders]
      security:
        - bearerAuth: []
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
          example: LOAD500
      responses:
        '202':
          description: Apex dispatched the load tender to FreightBridge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchLoadTenderResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/LoadNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessValidationFailure'
        '503':
          $ref: '#/components/responses/DependencyFailure'
  /v1/shipment-statuses:
    post:
      summary: Receive shipment status information
      description: Apex-owned endpoint that will eventually receive transformed shipment statuses from FreightBridge.
      operationId: receiveShipmentStatus
      tags: [Shipment Status]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApexShipmentStatus'
            examples:
              pickedUp:
                $ref: '#/components/examples/StatusPickedUp'
              inTransit:
                $ref: '#/components/examples/StatusInTransit'
              arrived:
                $ref: '#/components/examples/StatusArrived'
              delivered:
                $ref: '#/components/examples/StatusDelivered'
      responses:
        '202':
          description: Shipment status accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedPartnerEvent'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/LoadNotFound'
        '422':
          $ref: '#/components/responses/BusinessValidationFailure'
        '500':
          $ref: '#/components/responses/SystemFailure'
        '503':
          $ref: '#/components/responses/DependencyFailure'
  /v1/loads/{loadId}:
    get:
      summary: Retrieve a load by Apex load ID
      operationId: getLoadById
      tags: [Loads]
      security:
        - bearerAuth: []
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
          example: LOAD500
      responses:
        '200':
          description: Load found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApexLoad'
              examples:
                validLoad:
                  $ref: '#/components/examples/ValidLoadTender'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/LoadNotFound'
        '500':
          $ref: '#/components/responses/SystemFailure'
        '503':
          $ref: '#/components/responses/DependencyFailure'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: Bearer token placeholder for the MVP contract. No real token is included in this repository.
  schemas:
    ApexLoad:
      type: object
      required:
        - loadId
        - bolNumber
        - equipmentType
        - weightLbs
        - commodityDescription
        - pickup
        - delivery
        - createdAt
        - updatedAt
      properties:
        loadId:
          type: string
          example: LOAD500
        bolNumber:
          type: string
          example: BOL900
        purchaseOrderNumber:
          type: string
          example: PO111
        customerReference:
          type: string
          example: CUST-REF-500
        equipmentType:
          type: string
          enum: [VAN_53, REEFER_53, FLATBED]
          example: VAN_53
        weightLbs:
          type: number
          minimum: 1
          example: 42000
        pieces:
          type: integer
          minimum: 1
          example: 22
        commodityDescription:
          type: string
          example: Packaged auto parts
        pickup:
          $ref: '#/components/schemas/ApexLocation'
        delivery:
          $ref: '#/components/schemas/ApexLocation'
        references:
          type: array
          items:
            $ref: '#/components/schemas/ApexReference'
        createdAt:
          type: string
          format: date-time
          example: '2026-09-19T14:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-09-19T14:05:00Z'
    ApexLocation:
      type: object
      required: [facilityName, address1, city, state, postalCode, scheduledDateTime]
      properties:
        facilityName:
          type: string
          example: ABC Factory
        address1:
          type: string
          example: 200 Industrial Rd
        address2:
          type: string
          nullable: true
          example: Dock 4
        city:
          type: string
          example: Aurora
        state:
          type: string
          minLength: 2
          maxLength: 2
          example: IL
        postalCode:
          type: string
          example: '60505'
        scheduledDateTime:
          type: string
          format: date-time
          example: '2026-10-01T14:00:00Z'
    ApexReference:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [BOL, PO, CUSTOMER_REF, APPOINTMENT]
          example: CUSTOMER_REF
        value:
          type: string
          example: CUST-REF-500
        description:
          type: string
          example: Customer routing reference
    ApexTenderResponse:
      type: object
      required: [loadId, decision, carrierCode, decidedAt]
      properties:
        loadId:
          type: string
          example: LOAD500
        decision:
          type: string
          enum: [ACCEPTED, REJECTED]
          example: ACCEPTED
        carrierCode:
          type: string
          example: MWCX
        carrierLoadNumber:
          type: string
          example: MWC900500
        reasonCode:
          type: string
          example: CAPACITY_UNAVAILABLE
        message:
          type: string
          example: Tender accepted by Midwest Carrier.
        decidedAt:
          type: string
          format: date-time
          example: '2026-09-19T14:45:00Z'
    ApexShipmentStatus:
      type: object
      required: [loadId, carrierCode, statusCode, occurredAt]
      properties:
        loadId:
          type: string
          example: LOAD500
        carrierCode:
          type: string
          example: MWCX
        statusCode:
          type: string
          enum: [PICKED_UP, IN_TRANSIT, ARRIVED, DELIVERED]
          example: PICKED_UP
        statusDescription:
          type: string
          example: Shipment picked up at origin.
        occurredAt:
          type: string
          format: date-time
          example: '2026-10-01T14:30:00Z'
        city:
          type: string
          example: Aurora
        state:
          type: string
          example: IL
    AcceptedLoadTender:
      type: object
      required: [loadId, status, acceptedAt]
      properties:
        loadId:
          type: string
          example: LOAD500
        status:
          type: string
          example: ACCEPTED_FOR_PROCESSING
        acceptedAt:
          type: string
          format: date-time
          example: '2026-09-19T14:06:00Z'
    AcceptedPartnerEvent:
      type: object
      required: [eventId, status, acceptedAt]
      properties:
        eventId:
          type: string
          example: EVT-LOAD500-001
        status:
          type: string
          example: ACCEPTED
        acceptedAt:
          type: string
          format: date-time
          example: '2026-09-19T14:46:00Z'
    DispatchLoadTenderResponse:
      type: object
      required: [loadId, status, correlationId, freightbridge]
      properties:
        loadId:
          type: string
          example: LOAD500
        status:
          type: string
          example: DISPATCHED
        correlationId:
          type: string
          example: corr-LOAD500-001
        freightbridge:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: Request body is invalid.
            correlationId:
              type: string
              example: corr-LOAD500-001
  responses:
    InvalidRequest:
      description: Invalid request syntax or malformed JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authenticated client is not authorized for the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    LoadNotFound:
      description: Load ID was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Duplicate or conflicting request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BusinessValidationFailure:
      description: Business or schema validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    SystemFailure:
      description: Apex system failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    DependencyFailure:
      description: Apex dependency failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  examples:
    ValidLoadTender:
      value:
        loadId: LOAD500
        bolNumber: BOL900
        purchaseOrderNumber: PO111
        customerReference: CUST-REF-500
        equipmentType: VAN_53
        weightLbs: 42000
        pieces: 22
        commodityDescription: Packaged auto parts
        pickup:
          facilityName: ABC Factory
          address1: 200 Industrial Rd
          address2: Dock 4
          city: Aurora
          state: IL
          postalCode: '60505'
          scheduledDateTime: '2026-10-01T14:00:00Z'
        delivery:
          facilityName: XYZ Warehouse
          address1: 900 Commerce St
          address2: null
          city: Detroit
          state: MI
          postalCode: '48201'
          scheduledDateTime: '2026-10-02T18:00:00Z'
        references:
          - type: CUSTOMER_REF
            value: CUST-REF-500
            description: Customer routing reference
        createdAt: '2026-09-19T14:00:00Z'
        updatedAt: '2026-09-19T14:05:00Z'
    TenderAccepted:
      value:
        loadId: LOAD500
        decision: ACCEPTED
        carrierCode: MWCX
        carrierLoadNumber: MWC900500
        message: Tender accepted by Midwest Carrier.
        decidedAt: '2026-09-19T14:45:00Z'
    TenderRejected:
      value:
        loadId: LOAD500
        decision: REJECTED
        carrierCode: MWCX
        reasonCode: CAPACITY_UNAVAILABLE
        message: No available driver for requested pickup window.
        decidedAt: '2026-09-19T14:45:00Z'
    StatusPickedUp:
      value:
        loadId: LOAD500
        carrierCode: MWCX
        statusCode: PICKED_UP
        statusDescription: Shipment picked up at origin.
        occurredAt: '2026-10-01T14:30:00Z'
        city: Aurora
        state: IL
    StatusInTransit:
      value:
        loadId: LOAD500
        carrierCode: MWCX
        statusCode: IN_TRANSIT
        statusDescription: Shipment is en route to destination.
        occurredAt: '2026-10-01T20:15:00Z'
        city: Toledo
        state: OH
    StatusArrived:
      value:
        loadId: LOAD500
        carrierCode: MWCX
        statusCode: ARRIVED
        statusDescription: Shipment arrived at destination.
        occurredAt: '2026-10-02T17:45:00Z'
        city: Detroit
        state: MI
    StatusDelivered:
      value:
        loadId: LOAD500
        carrierCode: MWCX
        statusCode: DELIVERED
        statusDescription: Shipment delivered to destination.
        occurredAt: '2026-10-02T18:15:00Z'
        city: Detroit
        state: MI